make security cooperation permission

+++ app/Http/Controllers/Admin/CooperationController.php
@@ -4,10 +4,12 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\CooperationRequest;
+use App\Models\Cooperation;
 use App\Services\CooperationService;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Storage;
 
@@ -22,6 +24,9 @@ class CooperationController extends Controller
 
     public function store(CooperationRequest $request): RedirectResponse
     {
+        // check permission
+        Gate::authorize('create', Cooperation::class);
+
         $credentials = $request->validated();
 
         if ($request->file('image')) {
@@ -38,6 +43,10 @@ class CooperationController extends Controller
 
     public function update(int $id, CooperationRequest $request): RedirectResponse
     {
+        // check permission
+        $coopeartion = Cooperation::where('id', $id)->firstOrFail();
+        Gate::authorize('update', $coopeartion);
+
         $credentials = $request->validated();
 
         if ($request->file('image')) {
@@ -58,6 +67,10 @@ class CooperationController extends Controller
 
     public function destroy(int $id): RedirectResponse
     {
+        // check permission
+        $coopeartion = Cooperation::where('id', $id)->firstOrFail();
+        Gate::authorize('delete', $coopeartion);
+
         DB::transaction(function () use ($id) {
             $cooperation = $this->cooperationService->delete($id);
             Log::info($cooperation);
